rename(office-studio): rename Office Suite to Office Studio
Renames the app id, display name, component, view directory, and store
catalog entry from Office Suite to Office Studio across the frontend and
backend allowlist/version/trust maps. Drops the unimplemented Data rail
item from the app's left rail. The /api/office/* routes and the office doc
kind values (write/calc/db/slides) are unchanged, they are storage kinds,
not the app id.
