+++ desktop/src/apps/OfficeStudioApp.test.tsx
@@ -1,27 +1,26 @@
 import { render, screen, fireEvent } from "@testing-library/react";
 import { describe, it, expect } from "vitest";
-import { OfficeSuiteApp } from "./OfficeSuiteApp";
+import { OfficeStudioApp } from "./OfficeStudioApp";
 
 function renderApp() {
-  return render(<OfficeSuiteApp windowId="test-window" />);
+  return render(<OfficeStudioApp windowId="test-window" />);
 }
 
-describe("OfficeSuiteApp", () => {
+describe("OfficeStudioApp", () => {
   it("renders all rail items", () => {
     renderApp();
-    const nav = screen.getByRole("navigation", { name: "Office Suite views" });
+    const nav = screen.getByRole("navigation", { name: "Office Studio views" });
     expect(nav).toBeDefined();
     // query within nav to avoid ambiguity with toolbar buttons
     expect(nav.querySelector('[aria-label="Write"]')).toBeTruthy();
     expect(nav.querySelector('[aria-label="Calc"]')).toBeTruthy();
     expect(nav.querySelector('[aria-label="Slides"]')).toBeTruthy();
-    expect(nav.querySelector('[aria-label="Data"]')).toBeTruthy();
     expect(nav.querySelector('[aria-label="Assist"]')).toBeTruthy();
   });
 
   it("shows Write view by default with Write rail item active", () => {
     renderApp();
-    const nav = screen.getByRole("navigation", { name: "Office Suite views" });
+    const nav = screen.getByRole("navigation", { name: "Office Studio views" });
     const writeBtn = nav.querySelector('[aria-label="Write"]') as HTMLElement;
     expect(writeBtn).toBeTruthy();
     expect(writeBtn.getAttribute("aria-current")).toBe("page");
