+++ desktop/src/apps/officestudio/CalcView.tsx
