+++ desktop/src/apps/officestudio/SlidesView.tsx
